TMS: Fixed request objects having Integer instead of Long for ids

+++ tms-api/src/main/java/pu/master/tmsapi/models/requests/CommentRequest.java
@@ -12,9 +12,9 @@ public class CommentRequest
 
     private LocalDateTime timePosted;
 
-    private Integer author;
+    private Long author;
 
-    private Integer task;
+    private Long task;
 
 
     public CommentRequest()
@@ -34,13 +34,13 @@ public class CommentRequest
     }
 
 
-    public Integer getAuthor()
+    public Long getAuthor()
     {
         return author;
     }
 
 
-    public Integer getTask()
+    public Long getTask()
     {
         return task;
     }
